+++ services/preview/hardwaresecuritymodules/mgmt/2018-10-31-preview/hardwaresecuritymodules/CHANGELOG.md
@@ -1,2 +1,32 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Removed Funcs
+
+1. PossibleNameValues() []Name
+
+### Signature Changes
+
+#### Const Types
+
+1. SafeNetLunaNetworkHSMA790 changed type from Name to SkuName
+
+#### Struct Fields
+
+1. Sku.Name changed type from Name to SkuName
+
+## Additive Changes
+
+### New Constants
+
+1. SkuName.PayShield10KLMK1CPS250
+1. SkuName.PayShield10KLMK1CPS2500
+1. SkuName.PayShield10KLMK1CPS60
+1. SkuName.PayShield10KLMK2CPS250
+1. SkuName.PayShield10KLMK2CPS2500
+1. SkuName.PayShield10KLMK2CPS60
+
+### New Funcs
+
+1. PossibleSkuNameValues() []SkuName